Reduce size of the opamp_rev array
git-svn-id: https://svn.code.sf.net/p/sidplay-residfp/code/trunk/libsidplayfp@1501 75748be7-0387-4326-94c7-f41ce0883d2c

+++ builders/residfp-builder/residfp/FilterModelConfig.h
@@ -95,7 +95,7 @@ private:
     //@}
 
     /// Reverse op-amp transfer function.
-    int opamp_rev[1 << 16];
+    unsigned short opamp_rev[1 << 16];
 
 private:
     double getDacZero(double adjustment) const { return dac_zero - (adjustment - 0.5) * 2.; }
